[+BUGFIX] Roundtrip: Fix class and method modifiers
With roundtrip enabled, a single space is added before each class declaration.
This changes the class builder to not add this space by fixing the retrieval
of class and method modifiers. Those are resolved with Reflection::getModifierNames()
which returns an array of modifiers. Before this change, the first one (or NULL
if no modifiers are set) was added to the result array. The NULL value in the
array resulted in the space before the class definition. This changes this so the
result is directly used (i.e. array of all modifiers or NULL) is used. Then the
modifier NULL is not added before class definitions avoiding the space and for
example abstract methods have both modifiers (abstract and public) set and not
just one.

Change-Id: I8ad6cd68d5e8a32a2fc7f7e802a3153a61609510