[BUGFIX] Fix generated TCA for Single-Table Inheritance
we need to use require() instead of require_once() to include the subclass
TCA reliably. I just had the case that in some circumstances, the TCA
is loaded multiple times (for example when having multiple plugins
on the same page). For the first time, the TCA is complete, i.e. contains
both the parent and the sub-class fields. However, at the second load,
only the parent class is loaded because require_once skips the subclasses
as PHP had them already loaded.

That's a really hard-to-track-down issue which we should solve in the
Extension Builder.

Change-Id: I30bdbb13a14669e4a70af291157c15da67287915