[TASK] Compatibility with 6.0 Part 1(!)
All tests are green again. Removed all usage of deprecated class names in extension_builder.

This is called part 1 since the generated extensions are not yet namespaced!!
This will happen in part 2. This revision's edit mode should be compatible with 4.7 generated extensions!
The next revision will generate namespaced extensions and the edit mode will be only applicable for
extensions generated with that revision (or later)

Change-Id: I82b1aabe177f6fa0361cecf5d7924a417186d704