[TASK] TYPO3v6.0 compatibility Part 2
Now the generated extensions are namespaced! Also a vendor name is added and can be modified in the GUI.
Edit (roundtrip) mode is NOT supported yet. (So there will be a part 3 of compatibility update)

Somehow the generated plugins can create,delete, list and show models, but not update.
I guess that's a extbase bug, since the repository update does not result in updated models.

Change-Id: I9339ded78dc8347bfb5d74b805f30bbfb4bd0ba3