[BUGFIX] Make Extension Builder work with current master
In the current TYPO3 master branch (6.2) all modules
require a module token to avoid CSRF attacks.

Because of that, we need to change the smd endpoint
URL to be dynamic to generate the target URL
properly.

The change will make it work in master branch
but is still compatible with older TYPO3 versions.

Resolves: #56465
Change-Id: I866e0756ba134601ec53c8752495e1e8622fe6bc
Reviewed-on: https://review.typo3.org/27970
Reviewed-by: Steffen Müller
Tested-by: Steffen Müller
Reviewed-by: Nico de Haen
Tested-by: Nico de Haen