[TASK] Enable FAL related property types
The FAL integration in extbase is still quite incomplete.
This patch enables adding files or images as properties.

A property of type file or image is in fact a relation.
But if you create a relation to sys_file_reference in the
modeler, you won't get the appropriate form element for
adding a reference or uploading files in the Backend.

So this patch adds a property of type file_reference if
maxItems is 1 and a 1:n relation (ObjectStorage),
if maxItems is > 1.

The folder property, is dropped since it is not
supportd by FAL. Instead a relation to sys_file_collection
could be added as usual.

Releases: 6.2
Resolves: #57791
Change-Id: Ia0b7f3aa08b2a9055b150fbb37b65164c98a58e4
Reviewed-on: https://review.typo3.org/29447
Reviewed-by: Nico de Haen
Tested-by: Nico de Haen