[TASK] Enable merging of xliff files
This patch enables "merging" of xliff localization files.
Manually added key/label pairs are preserved on rewriting
the files, but changes made in the modeler are applied.

When renaming a model or a property the old id/label is
not removed, since we would have to load all labels from
the previous version of the extension to identify renamed
items.

Support for xml format is dropped.

Releases: 6.2
Resolves: #57688
Change-Id: I003a95576d298c2127644b82a087d738ca095456
Reviewed-on: https://review.typo3.org/29527
Reviewed-by: Nico de Haen
Tested-by: Nico de Haen