[DOCS] Use spelling "domain model" and "domain objects"
There are many notations for the description of domain objects of
domain-driven design: the only valid one in this documentation is
"domain object" and the domain model (singular) is the sum of all
domain objects (plural).

The term "domain model" was incorrectly used for domain objects because
the classes of the domain object types "entity" and "value object" are
managed in a special folder /Classes/Domain/Model/ that reads as
"domain model".

Use the term "object" instead of "domain object" only when it is clear
from the context that we are talking about domain objects - and not
class objects, for example.

The naming conflict that we usually only talk about entities and value
objects when we talk about domain objects remains unresolved: In
domain-driven design, repositories, factories, services, and events are
also domain objects. However, as long as there is no separate name for
entities and value objects, we refer to these two exclusively as
"domain objects" in most lines.