[FEATURE] Set storage path in extension configuration
In addition to automatic detection of the storage path, the Extension
Builder now supports explicit specification of the storage path via
the new extension configuration parameter `storageDir`. It helps to
cope with scenarios where the save path cannot be determined
automatically, e.g. in test environments in Composer mode.

This corresponds to the setting of the backup path.