[BUGFIX] Language file not merged (#605)
The language files inside Resources/Private/Language were not merged, after saving the extension inside the extension builder.
It happened because the suffix was added twice.

The table name was added to the filename so it caused a filename like "tx_tablename_tca.xlf_tablename_tca" and this file didn't exist.