[TASK] Use current standard for web-dir
To make contribution and maintainance of configuration, scripts etc.
for testing easier, we try to use a common standard across TYPO3
extensions.

The current standard is:

- web-dir: .Build/Web
- bin-dir: .Build/bin
- vendor-dir: .Build/vendor

This is used in the official example extensions (tea, docs-examples)
as well as in most of the checked community extensions (in a subset
of 20 of the most popular v12 extensions).

As extension_builder is very influential in this area and is used to
create extensions from scratch, it is useful to stick to the standard
in this extension as well.

Related: TYPO3-Documentation/tea#802